Moved alert to a more generic place
